# Adding OpenMP support in multi_corr.c