@chrisdjscott 2017-08-17
Initial benchmarking results comparing the OpenMP version with the Python multiprocessing version (both call the same underlying C function: `normxcorr_fftw_main`). These benchmarks were with the real data and 100 templates.

![results_100templates](https://user-images.githubusercontent.com/6309552/29394776-badd3e3e-8360-11e7-80ba-cb0b1717cf6e.png)

Memory usage (top graph):
* Peak memory usage increases with core count and is slightly lower for OpenMP
* The number of cores is limited by memory and the number of channels (9 in this case)

Run time (bottom graph):
* OpenMP much faster on 1 core but Python catches up as the number of cores increases
* This is because more code is parallelised in the Python version (see below)
* OpenMP slightly faster at 9 (= num_channels) cores

![results_py_vs_c](https://user-images.githubusercontent.com/6309552/29394924-b4a2ff8a-8361-11e7-9136-b17309019800.png)

* In the OpenMP version there is some serial stuff done in Python before and after the call to C (see `utils.correlate.fftw_multi_normxcorr`)
* This gets run on each thread (i.e. in parallel) in the multiprocessing version (see `utils.correlate.fftw_normxcorr`)
* As we increase cores the serial Python code becomes a bottleneck in the openmp version

Next I plan to look at:

* Adding threading in the main C routine that computes the cross correlation (`normxcorr_fftw_main`)
  - should allow us to utilise more cores
  - increasing cores here should not increase memory usage
  - will need to do some benchmarking to see if it is worthwhile
* Look at the serial Python code in the OpenMP version to see if it can be optimised or is worth running in parallel