@chrisdjscott 2017-08-29
I have modified the Python function to move the parts that were taking significant time into C, so it can run in parallel (C should be faster too). I profiled that function using [line_profiler](https://pypi.python.org/pypi/line_profiler) first, see below if interested. The "% Time" column is important. Note only 57.5% in the C function (some lines with very low "% Time" omitted).

```
Function: fftw_multi_normxcorr at line 342

Line        Hits         Time  Per Hit   % Time  Line Contents
==============================================================
...
   399         1            8      8.0      0.0      cccs = np.empty((n_channels, n_templates, image_len - template_len + 1),
   400         1      6440302 6440302.0    8.3             np.float32).flatten(order='C')
   401         1            6      6.0      0.0      ret = utilslib.multi_normxcorr_fftw(
   402         1            2      2.0      0.0          template_array, n_templates, template_len, n_channels, stream_array,
   403         1     44357571 44357571.0   57.5          image_len, cccs, fft_len)
...
   408        10           21      2.1      0.0      for j in range(n_channels):
   409       909         2002      2.2      0.0          for i in range(n_templates):
   410       900         2129      2.4      0.0              if not used_chans[j][i]:
   411       593      4266116   7194.1      5.5                  cccs[j][i] = np.zeros(image_len - template_len + 1)
   412         1      2517897 2517897.0     3.3      cccs[np.isnan(cccs)] = 0.0
   413         1      8582607 8582607.0     11.1      if np.any(np.abs(cccs) > 1.01):
   414                                                   print('Normalisation error in C code')
   415                                                   print(cccs.max())
   416                                                   print(cccs.min())
   417                                                   raise MemoryError()
   418         1      2552231 2552231.0      3.3      cccs[cccs > 1.0] = 1.0
   419         1      2518577 2518577.0      3.3      cccs[cccs < -1.0] = -1.0
   420        10           40      4.0      0.0      for j, seed_id in enumerate(seed_ids):
   421       909         4753      5.2      0.0          for i in range(len(pad_array[seed_id])):
   422       900         3602      4.0      0.0              cccs[j][i] = np.append(
   423       900         4176      4.6      0.0                  cccs[j][i],
   424       900      5841140   6490.2      7.6                  np.zeros(pad_array[seed_id][i]))[pad_array[seed_id][i]:]
   425         1           13     13.0      0.0      cccs = cccs.reshape(n_channels, n_templates, image_len - template_len + 1)
   426         1            2      2.0      0.0      return cccs, used_chans
```

We can remove the `flatten` on line 400 since the array is guaranteed to be `C_CONTIGUOUS` - when you access the underlying data in C it will be implicitly flattened. The array in this line was very big and `flatten` creates a full copy of the array, so that's probably why it was taking so long. The other time consuming lines can be moved to C. 

After doing this we get the following updated graph (OMP2 is the new version).

![results_100templates_omp2](https://user-images.githubusercontent.com/6309552/29798497-6719003a-8cb1-11e7-894d-701c9fda007b.png)

- Memory usage is similar
- Run time is significantly lower with OMP2
  * Multiprocessing with 9 threads takes 89.7 seconds
  * Original OpenMP with 9 threads takes 80.4 seconds
  * New OpenMP with 9 threads takes 51.2 seconds (1.8x faster than multiprocessing and 1.6x faster than the original OpenMP version)

I still want to look at threading in the main C routine and maybe run some bigger tests on this version.