@calum-chamberlain 2017-08-31
Just had a thought: In the python function that calls the python function that calls the c-function (convoluted) within correlate.py (multi_normxcorr.py?) we are summing the results of the c-function as soon as we get them back.  It would be more memory efficient to compute the sums in the c-function, as long as it doesn't impact parallel computation. Do you think that would be feasible? We don't need the cross-correlations for individual channels, just the sum of all channels for each template.
What do you think @chrisdjscott ?