@chrisdjscott 2017-08-31
Yes, sounds like it might make sense to do this in the C code too @calum-chamberlain - I'll have a look into it.