@chrisdjscott 2017-09-05
@calum-chamberlain - I moved the sum into C and we got a small boost in performance - 47.0 seconds down from 51.2s (OpenMP) and 89.7s (multiprocessing), so about 1.9x faster than multiprocessing now. Memory usage doesn't change much as I think the full array is still needed for the calculation.

I'll try merging the changes from develop into here to get rid of the conflict (and see if any change in performance). Then I will look to see if threading in `normxcorr_fftw_main` will help. Does that sound OK?