@calum-chamberlain 2017-09-05
Nice, okay - I was thinking that the summing could go within `normxcorr_fftw_main` - to replace the full array of cross-correlation (rather than assigning each value, each value could be added to the previous one), which might have reduced memory cost.  I guess that might not work calling in parallel and potentially trying to sum multiple times to the same bit of memory?