@chrisdjscott 2017-09-05
We would have to use atomic operations to do that I think @calum-chamberlain, which could slow things down. I'll have a look to see how easy it will be to implement as it would be good to get an idea of the performance cost. It looks like about 12GB memory is used in that big array (in the test case you sent me with 9 channels and 100 templates), while only ~1.3GB needed after the sums so could save a reasonable amount.