@calum-chamberlain 2017-09-05
That is a concept far beyond my grasp of computing...